requirements.txt
flask
ai4bharat-transliteration

